+++ apps/web/src/lib/format-time.ts
@@ -1,0 +1,15 @@
+export function formatRelativeTime(date: Date): string {
+  const now = Date.now();
+  const diff = now - new Date(date).getTime();
+  const seconds = Math.floor(diff / 1000);
+  if (seconds < 60) return 'just now';
+  const minutes = Math.floor(seconds / 60);
+  if (minutes < 60) return `${minutes} min. ago`;
+  const hours = Math.floor(minutes / 60);
+  if (hours < 24) return `${hours} hr. ago`;
+  const days = Math.floor(hours / 24);
+  if (days < 30) return `${days}d ago`;
+  return new Intl.DateTimeFormat(undefined, { month: 'short', day: 'numeric' }).format(
+    new Date(date),
+  );
+}
