+++ apps/web/src/pages/editor/components/comment-bubble.tsx
@@ -1,5 +1,6 @@
 import { useMemo, useState } from 'react';
 import type { CommentResponse } from '@draftila/shared';
+import { formatRelativeTime } from '@/lib/format-time';
 
 interface CommentBubbleProps {
   x: number;
@@ -26,22 +27,6 @@ function hasUnread(thread: CommentResponse | null): boolean {
     }
   }
   return false;
-}
-
-function formatRelativeTime(date: Date): string {
-  const now = Date.now();
-  const diff = now - new Date(date).getTime();
-  const seconds = Math.floor(diff / 1000);
-  if (seconds < 60) return 'just now';
-  const minutes = Math.floor(seconds / 60);
-  if (minutes < 60) return `${minutes} min. ago`;
-  const hours = Math.floor(minutes / 60);
-  if (hours < 24) return `${hours} hr. ago`;
-  const days = Math.floor(hours / 24);
-  if (days < 30) return `${days}d ago`;
-  return new Intl.DateTimeFormat(undefined, { month: 'short', day: 'numeric' }).format(
-    new Date(date),
-  );
 }
 
 export function CommentBubble({
